% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{convert_colo}
\alias{convert_colo}
\title{Convert shorthand colour name to a colour vector}
\usage{
convert_colo(colo)
}
\arguments{
\item{colo}{\code{NULL}, a character vector of colour hex codes, or a single
character keyword. If it is a character of length 1 it is expanded into a
character vector of colours (see Details). Otherwise it is returned
unchanged.}
}
\value{
A character vector of hex colour codes, or \code{NULL} if \code{colo} is
\code{NULL}.
}
\description{
\code{convert_colo} accepts \code{NULL}, a character vector of colour hex codes, or a
single keyword and returns a character vector of hex colours (or \code{NULL}).
This helper is intended for normalising user-supplied colour arguments into
a vector usable by \code{ggplot2::scale_fill_manual()} and similar functions.
Tidies a situation where a variable (usually called \code{colo}) is passed around
as either a string of hex code colours or a shortcut name/keyword.
Mainly kept for backwards-compatibility - in most contexts this function
SHOULD NOT BE USED, as \code{OMESurvey::get_OME_colours()} is more direct &
transparent and less hard-coded.
}
\details{
The following single-string keywords are recognised and expanded:
\describe{
\item{\code{"likert5"}}{5-colour palette: \code{c("#009BC1","#08607E","#10263B","#732C53","#D7336C")}}
\item{\code{"scale4"}}{4-colour palette: \code{c("#009BC1","#08607E","#732C53","#D7336C")}}
\item{\code{"ynsn3"}}{3-colour palette: \code{c("#009BC1","#10263B","#D7336C")}}
\item{\code{"scale2"}}{2-colour palette: \code{c("#009BC1","#D7336C")}}
}
Any other single string is returned as-is, i.e. it DOESN'T throw an error/warning!
}
\examples{
convert_colo(NULL)            # NULL -> NULL
convert_colo(c("#000000","#FFFFFF"))  # vector -> returned unchanged
convert_colo("likert5")       # returns the 5-colour likert palette
convert_colo("an_odd_name")       # Unintended use: unknown and so returned unchanged
convert_colo(c("thing","amijig"))  # Unintended use: vector -> returned unchanged

}
