% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_NA}
\alias{convert_NA}
\title{Convert "N-A" to NA across a data frame}
\usage{
convert_NA(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Convert "N-A" to NA across a data frame
}
\details{
Loops over all columns and changes "N-A" to NA.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
OMESurvey::survey_example |> convert_NA()

\dontshow{\}) # examplesIf}
}
