% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_initial_analysis_report}
\alias{create_initial_analysis_report}
\title{Initial analysis report}
\usage{
create_initial_analysis_report(
  survey_data,
  survey_name,
  participant_ID = "Barcode_ID",
  report_kind = "interactive",
  demographic_columns = NULL,
  output_dir = NULL,
  output_file = NULL,
  reference_docx = NULL
)
}
\arguments{
\item{survey_data}{The cleaned survey data}

\item{survey_name}{The name of the survey}

\item{participant_ID}{column name of the column in \code{survey_data} corresponding to the participant ID.}

\item{report_kind}{Either 'interactive'(.html) or 'static'(.docx) for the report type}

\item{demographic_columns}{column names (or indices) of the columns in \code{survey_data} corresponding to demographic columns. These will be used to compare missingness across the demographics.}

\item{output_dir}{The path of the output directory}

\item{output_file}{the filename of the output. If \code{NULL} we use paste0(survey_name, '_initial_analysis_report', "extension") as the filename.}

\item{reference_docx}{Path to a .docx file whose style (design) the report copies.}
}
\value{
html/word report
}
\description{
Create a report on the initial finding of a survey.
}
