% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_check.R
\name{factor_survey_data}
\alias{factor_survey_data}
\title{Factor survey questions in data}
\usage{
factor_survey_data(data, survey_values = OMESurvey::survey_values, verbose = T)
}
\arguments{
\item{data}{survey data}

\item{survey_values}{a list of allowed grouped survey values. Default is OMESurvey::survey_values.}

\item{verbose}{Flag (TRUE/FALSE) for whether to print console messages.}
}
\value{
data with factored survey columns
}
\description{
Factor survey questions in data
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
data = OMESurvey::survey_example
for(i in 1:ncol(data)){
  data[[i]][which(data[[i]] == 'N-A')] = NA
}
clean = data |> factor_survey_data()
\dontshow{\}) # examplesIf}
}
