% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{get_OME_colours}
\alias{get_OME_colours}
\title{get colours for the OME}
\usage{
get_OME_colours(n, type = "contrast")
}
\arguments{
\item{n}{the number of colours.}

\item{type}{either "contrast", "complementary" or "distinct".}
}
\value{
a vector of "n" colours.
}
\description{
get colours for the OME
}
\details{
Base colours for palettes:
\itemize{
\item OME complementary colours = "#10263B" (Dark Navy/Nottingham Blue), "#009BC1" (Blue/Malaysia Sky Blue) and "#37B4B0"(Turquoise/Trent Turquoise).
\item OME contrast colours = "#009BC1" (Blue/Malaysia Sky Blue), "#10263B" (Dark Navy/Nottingham Blue) and '#D7336C' (Pink/Pioneering Pink).
\item OME distinct colours = "#009BC1"(Blue/Malaysia Sky Blue),'#DEB406' (Gold/Rebel's Gold), '#D7336C' (Pink/Pioneering Pink) and '#10263B' (Dark Navy/Nottingham Blue).
}

If more than the designated colours are specified we interpolate the colours using \code{grDevices::colorRampPalette()}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_OME_colours(3) |> scales::show_col()
get_OME_colours(3, type = 'complementary') |> scales::show_col()
get_OME_colours(9, type = 'distinct') |> scales::show_col()
\dontshow{\}) # examplesIf}
}
