% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theme_and_questions}
\alias{get_theme_and_questions}
\title{Extract theme and questions from column names}
\usage{
get_theme_and_questions(theme_and_questions)
}
\arguments{
\item{theme_and_questions}{column names of a single theme and it's respective questions.}
}
\value{
a list of length 2, where
\itemize{
\item $theme string of the theme name
\item $questions vector of strings with the questions
}
}
\description{
Extract theme and questions from column names
}
