% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theme_columns}
\alias{get_theme_columns}
\title{Retrieve theme columns}
\usage{
get_theme_columns(data, theme)
}
\arguments{
\item{data}{survey data}

\item{theme}{A string containing the theme name (with or without "_" as spacing).}
}
\value{
the column indices of the questions relating to the theme
}
\description{
Retrieve theme columns
}
