% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{initial_bar}
\alias{initial_bar}
\title{Make stacked bar chart (possibly split into separate bars and facetted) in the OME style. (Dave's version.)}
\usage{
initial_bar(
  dat,
  percCut = NULL,
  colo = NULL,
  na.rm = FALSE,
  horiz = FALSE,
  text_scale = 1,
  fillLabText = NULL,
  xLabText = NULL,
  yLabText = "Proportion of responses",
  titleText = NULL,
  facet_labels = NULL,
  facet_layout = NULL
)
}
\arguments{
\item{dat}{data frame of 1, 2 or 3 variables. Bar charts show relative frequencies of the first variable,
split into separate bars according to the second variable (if present),
and facetted by the third variable (if present).
All variables should be factors (but characters might work?).}

\item{percCut}{numeric scalar (0-100). Cutoff below which percentages are not shown in bar segments. Default \code{5}.}

\item{colo}{(optional but recommended) vector of colours to use for the fill scale (character vector of colours,
length the same as the number of levels of the factor that is the first variable of \code{dat}.)
If \code{NULL} (default) the pallete from \code{OMESurvey::get_OME_colours(type='distinct')} is used.
When \code{na.rm = FALSE} a grey colour is prepended for the "No response" level.
Note that percentage labels are in white, so the pallete needs to work with that.
(... or this function needs upgrading!)}

\item{na.rm}{logical. If \code{TRUE} then remove \code{NA} responses from the data;
if \code{FALSE} (default) then they are converted to "(No response)" and treated as an additional allowed response.}

\item{horiz}{logical (default FALSE) determining whether to coord_flip() so that bars are horizontal and place legend below the plot.}

\item{text_scale}{positive number (default 1) for scaling the size of the percentage and #response labels.}

\item{fillLabText}{Optional character. Label for the fill/legend.
If \code{NULL} (the default) the legend title is removed (no label).
If \code{""} the name of the first variable in \code{dat} is used.}

\item{xLabText}{optional text for labelling the second-variable-in-\code{dat} axis.
If \code{NULL} (the default) the label is removed (no label).
If \code{""} the name of the first variable in \code{dat} is used.}

\item{yLabText}{optional text for labelling the 'proportion' axis. Default \code{"Proportion of responses"}.}

\item{titleText}{optional text for plot title. With default \code{NULL} the title is removed.}

\item{facet_labels}{Optional named character vector used as a labeller for
facets (names are original facet values, values are labels to display).
Default \code{NULL} uses the values of that variable.}

\item{facet_layout}{Optional character. If \code{"1row"} the facets are
arranged in a single row; otherwise default facet layout is used.
(\code{"1col"} option planned but not yet supported.)}
}
\value{
ggplot object - a stacked proportional bar chart.
}
\description{
\code{initial_bar} builds a proportional (stacked, filled) bar chart from a
tidy data frame of survey responses. The first column is treated as the
response factor (fill), the optional second column as the grouping/subdivision,
and the optional third column is used for faceting.
Percentages of bars are overlaid on the plot and number of observations/responses are shown at the end of each bar.
The function returns a \code{ggplot} object ready for further customization or direct printing.
}
\details{
Percentages are displayed in segments that comprise at least \code{percCut} of their bar;
the percentage displayed is this rounded to a whole number.
}
\examples{
# NEED A SHORT EXAMPLE
}
