% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{OME_boxplot}
\alias{OME_boxplot}
\title{Boxplot (horizontal) with optional right-side counts showing valid & total
#datapoints for boxes, in an OME style.}
\usage{
OME_boxplot(
  data,
  value_var,
  group_var,
  show_counts = TRUE,
  title = NULL,
  colour = OMESurvey::get_OME_colours(1),
  dashed_at = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{value_var}{Numeric variable to plot.}

\item{group_var}{Grouping variable (factor).}

\item{show_counts}{Logical; whether to display (valid/total) counts on the
right-hand axis. Defaults to TRUE.}

\item{title}{Optional plot title.}

\item{colour}{Boxplot outline colour (any valid R colour). Defaults to \href{1}{get_OME_colours}.}

\item{dashed_at}{Optional horizontal reference line.}
}
\value{
A ggplot object
}
\description{
Boxplot (horizontal) with optional right-side counts showing valid & total
#datapoints for boxes, in an OME style.
}
\note{
Future/possible extensions include support for dodging/colour
(and maybe vertical-ness and/or faceting).
}
\examples{

df <- tibble::tibble(
  group = factor(c("A", "A", "B", "B")),
  value = c(3.2, 4.1, 5.0, NA)
)

OME_boxplot(
  data = df,
  value_var = value,
  group_var = group,
  title = "Example boxplot"
)
}
\author{
Dave Sirl
}
