% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMESurvey-package.R
\docType{package}
\name{OMESurvey-package}
\alias{OMESurvey}
\alias{OMESurvey-package}
\title{OMESurvey: Tools for analysing survey data for the Observatory of Mathematical Education}
\description{
Tools for analysing survey data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jake-powell.github.io/OMESurvey/}
}

}
\author{
\strong{Maintainer}: Jake Powell \email{Jake.Powell2@nottingham.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
