% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{plot_many_questions}
\alias{plot_many_questions}
\title{Deprecated: Make a stacked bar chart summarising many survey questions}
\usage{
plot_many_questions(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{summary_plot_stacked_bar()}.}
}
\description{
This function is deprecated.
Use \code{\link{summary_plot_stacked_bar}} instead.
}
\seealso{
summary_plot_stacked_bar
}
