% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_theme}
\alias{plot_theme}
\title{plot survey theme result}
\usage{
plot_theme(
  data,
  theme = NA,
  theme_columns = NA,
  kind = "ggplot",
  rm99 = TRUE,
  survey_values = OMEStools::survey_values
)
}
\arguments{
\item{data}{survey data}

\item{theme}{A string containing the theme name (with or without "_" as spacing).}

\item{theme_columns}{column index relating to theme}

\item{kind}{what kind of object we want to be returned. Allowable values are "ggplot", "plotly" or "data.frame".}

\item{rm99}{Flag (TRUE/FALSE) for whether we want to set '99' values to NA to suppress multiple values as an option. Default is TRUE.}

\item{survey_values}{survey_values}
}
\value{
either a plotly object, ggplot object or a dataframe depending on \code{kind}.
}
\description{
plot survey theme result
}
