% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_theme_by_demographic}
\alias{plot_theme_by_demographic}
\title{Plot survey theme result by demographic}
\usage{
plot_theme_by_demographic(
  data,
  theme = NA,
  theme_columns = NA,
  demographic_column,
  agree_values = c("Agree a little", "Agree a lot"),
  kind = "ggplot",
  rm99 = TRUE
)
}
\arguments{
\item{data}{survey data}

\item{theme}{A string containing the theme name (with or without "_" as spacing).}

\item{theme_columns}{column index relating to theme}

\item{demographic_column}{the index of the demographic column}

\item{agree_values}{The answers that correspond to agree.}

\item{kind}{what kind of object we want to be returned. Allowable values are "ggplot", "plotly" or "data.frame".}

\item{rm99}{Flag (TRUE/FALSE) for whether we want to set '99' values to NA to suppress multiple values as an option. Default is TRUE.}
}
\value{
either a plotly object, ggplot object or a data frame depending on \code{kind}.
}
\description{
See \verb{Create survey figures} article for examples.
}
