% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{scale_colour_OME}
\alias{scale_colour_OME}
\alias{scale_color_OME}
\alias{scale_fill_OME}
\title{OME colour scales for ggplot2}
\usage{
scale_colour_OME(type = "distinct", name = NULL, ...)

scale_color_OME(type = "distinct", name = NULL, ...)

scale_fill_OME(type = "distinct", name = NULL, ...)
}
\arguments{
\item{type}{Character string specifying which palette to use.
Passed to \code{\link{get_OME_colours}}; see that function for available options.
Default is \code{"distinct"}.}

\item{name}{Character string specifying the title of the legend for this aesthetic.
Default is \code{NULL}, for no title.}

\item{...}{Additional arguments passed to \code{ggplot2::discrete}.}
}
\value{
A ggplot2 scale for colours.
}
\description{
Provides OME palettes for use with colour and fill aesthetics.
(Convenience wrapper around \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} using
the standard OME colour palette.)
}
\examples{
library(ggplot2)
library(dplyr)

# Examples here use scale_colour_OME(); scale_fill_OME() is exactly analogous.

# Prepare example dataset with labelled factors
mtcars_small <- mtcars |>
  mutate(
    am = factor(am, levels = c(0, 1), labels = c("Auto", "Man")),
    cyl = factor(cyl, levels = c(4, 6, 8))
  )

# Scatterplot using OME palette & theme for colour aesthetic
# (Not OME-like data and illustration would be better if the factor had more than two levels,
#  but hopefully illustrates use of scale_colour_OME() when colouring by a factor)
mtcars_small |>
  ggplot(aes(x = wt, y = mpg, colour = am)) +
  geom_point() +
  scale_colour_OME() +
  labs(
    x = "Weight (/1,000 lb)",
    y = "Miles per gallon",
    colour = "Transmission",
    title = "Fuel efficiency vs weight,\ncoloured by transmission type"
  )

# Use a sequential palette for an ordered factor ("divergent" is available too)
mtcars_small |>
  ggplot(aes(x = wt, y = mpg, colour = cyl)) +
  geom_point() +
  scale_colour_OME(type = "sequential") +
  labs(
    x = "Weight (/1,000 lb)",
    y = "Miles per gallon",
    colour = "Cylinders",
    title = "Fuel efficiency vs weight,\ncoloured by cylinders"
  )
}
\author{
Dave Sirl
}
