% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_OME_sharepoint_root}
\alias{set_OME_sharepoint_root}
\title{Explain how to set the OME SharePoint root folder}
\usage{
set_OME_sharepoint_root()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Prints instructions for setting the \code{OME_SHAREPOINT_ROOT} environment
variable in \code{.Renviron}. This should point to the user's local synced
SharePoint / OneDrive root folder, for example
\code{"C:/Users/yourID/OneDrive - Company"}.
}
\examples{
\dontrun{
set_OME_sharepoint_root()
}

}
