% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_example}
\alias{survey_example}
\title{Fake survey data}
\format{
An object of class \code{data.frame} with 484 rows and 66 columns.
}
\usage{
survey_example
}
\description{
Example survey data with 484 students answering 61 questions across 8 themes.
}
\details{
#' A data frame with 484 rows and 65 columns:
}
\keyword{datasets}
