% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_values}
\alias{survey_values}
\title{Survey values}
\format{
An object of class \code{list} of length 7.
}
\usage{
survey_values
}
\description{
A list of vectors where each vector contains strings representing survey answers to a single question. This includes:
\itemize{
\item 'Agree a lot', 'Agree a little', 'Neither', 'Disagree a little', 'Disagree a lot'
\item 'Almost always','Most of the time','Half the time','Some of the time','Almost never'
\item 'Often', 'Sometimes', 'Never'
\item 'A lot', 'Sometimes', 'Not much', 'Did not answer'
\item 'Good', 'Okay', 'Bad', 'Did not answer'
\item 'Like', 'Okay', "Don't like", 'Did not answer'
\item '1','2'
}
}
\details{
#' A list of length 3
}
\keyword{datasets}
