% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{theme_OME}
\alias{theme_OME}
\title{OME ggplot2 theme}
\usage{
theme_OME(base_size = 14, legend_placement = c("right", "bottom"))
}
\arguments{
\item{base_size}{Numeric. Base font size for the theme. Defaults to 14.}

\item{legend_placement}{Character. Controls position/orientation/etc of
legends. One of \code{"right"} or \code{"bottom"}. Defaults to \code{"right"}.}
}
\value{
A \code{ggplot2} theme object that can be added to a ggplot with \code{+}.
}
\description{
OME house style theme for \code{ggplot2}, based on \code{ggplot2::theme_bw()}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  p <- ggplot2::ggplot(mtcars, ggplot2::aes(x = factor(cyl))) +
    ggplot2::geom_bar() +
    ggplot2::labs(
      title = "A plot with non-OME-relevant data,",
      subtitle = "but which demos using theme_OME"
      )

  p + theme_OME()

  p + theme_OME(legend_placement = "bottom")

  # Set as the global default for the current session:
  ggplot2::theme_set(theme_OME())
}

}
\author{
Dave Sirl
}
