% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_logo}
\alias{add_logo}
\title{Add OME logo to a ggplot object}
\usage{
add_logo(p, type = "", position = "bottom right", logo_sizing = c(0.11, 0.2))
}
\arguments{
\item{p}{ggplot object}

\item{type}{Specifies the type on logo. Either 'bw' or 'standard'.}

\item{position}{Specifies the logo position. Either 'top left', 'top right', 'bottom left' or 'bottom right'}

\item{logo_sizing}{A numeric vector of length 2. The first element defines the amount of vertical spacing the logo uses (default is 11\% of the figure height). The second element defines the logo width (min 0, max 1 default is 0.2)}
}
\value{
original plot with the logo added.
}
\description{
Add OME logo to a ggplot object
}
\details{
This function currently always adds the logo outside the original plot (In an extended margin).

Note that you cannot use '+' to add this to the ggplot object, as the function returns a non-ggplot object (as we use gridExtra::grid.arrange() to add the logo). This should always be the last step of creating a plot.

PLay around with the logo sizing parameter until you find your desired logo size.
}
