% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{get_OME_colours}
\alias{get_OME_colours}
\title{get colours for the OME}
\usage{
get_OME_colours(n, type = "complementary")
}
\arguments{
\item{n}{the number of colours}

\item{type}{either "complementary" or "constrast"}
}
\value{
a vector of colours
}
\description{
get colours for the OME
}
