% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{get_OME_colours}
\alias{get_OME_colours}
\title{get colours for the OME}
\usage{
get_OME_colours(n, type = "contrast")
}
\arguments{
\item{n}{the number of colours.}

\item{type}{either "contrast/divergent", "complementary/sequential" or "distinct/qualitative".}
}
\value{
a vector of "n" colours.
}
\description{
get colours for the OME
}
\details{
Base colours for palettes:
\itemize{
\item OME complementary/sequential colours = "#10263B" (Dark Navy/Nottingham Blue), "#009BC1" (Blue/Malaysia Sky Blue) and "#37B4B0"(Turquoise/Trent Turquoise).
\item OME contrast/divergent colours = "#009BC1" (Blue/Malaysia Sky Blue), "#10263B" (Dark Navy/Nottingham Blue) and '#D7336C' (Pink/Pioneering Pink).
\item OME distinct/qualitative colours = "#009BC1"(Blue/Malaysia Sky Blue),'#DEB406' (Gold/Rebel's Gold), '#D7336C' (Pink/Pioneering Pink) and '#10263B' (Dark Navy/Nottingham Blue).
}

For 'contrast' and 'distinct' specific colour orderings are chosen for n <= 9, whereas for 'complementary' specific colours selected for n <= 3. If n is larger than we interpolate the palette using grDevices::colorRampPalette().
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
get_OME_colours(3) |> scales::show_col()
get_OME_colours(3, type = 'complementary') |> scales::show_col()
get_OME_colours(9, type = 'distinct') |> scales::show_col()
\dontshow{\}) # examplesIf}
}
