% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{get_OME_colours}
\alias{get_OME_colours}
\title{get colours for the OME}
\usage{
get_OME_colours(n, type = "complementary")
}
\arguments{
\item{n}{the number of colours.}

\item{type}{either "complementary" or "contrast".}
}
\value{
a vector of colours.
}
\description{
get colours for the OME
}
\details{
OME complementary colours = c("#10263A","#009BC2", "#37B4B0").
OME contrast colours = c("#009BC2", "#10263A", '#d7386b').

If more than 3 colours are specified we interpolate the colours using \code{grDevices::colorRampPalette()}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_OME_colours(3) |> scales::show_col()
get_OME_colours(3, type = 'contrast') |> scales::show_col()
get_OME_colours(10, type = 'contrast') |> scales::show_col()
\dontshow{\}) # examplesIf}
}
