% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{plot_many_questions}
\alias{plot_many_questions}
\title{Make a stacked bar chart summarising many survey questions}
\usage{
plot_many_questions(
  dat,
  labels_vec,
  percCut = 2,
  colo = NULL,
  order_values = NULL,
  titleText = NULL
)
}
\arguments{
\item{dat}{INCOMPLETE}

\item{labels_vec}{INCOMPLETE}

\item{percCut}{INCOMPLETE}

\item{colo}{INCOMPLETE}

\item{order_values}{INCOMPLETE}

\item{titleText}{INCOMPLETE}
}
\value{
INCOMPLETE
}
\description{
Longer version - INCOMPLETE
}
\examples{
# NEED A SHORT EXAMPLE HERE

}
