% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary_functions.R
\name{theme_OME}
\alias{theme_OME}
\title{ROME ggplot2 theme}
\usage{
theme_OME(
  base_size = 16,
  base_family = choose_font_family("Arial"),
  legend_placement = c("right", "bottom")
)
}
\arguments{
\item{base_size}{Numeric. Base font size for the theme. Defaults to 16.}

\item{base_family}{Base font family for plot text. Defaults to Arial when
available via \code{systemfonts}; otherwise falls back to \code{"sans"}.}
}
\value{
A \code{ggplot2} theme object that can be added to a ggplot with \code{+}.
}
\description{
An OME ggplot2 theme (based on \code{ggplot2::theme_bw()}).
}
\details{
!!! CURRENTLY COPIED FROM JAKE'S ROME_ggtheme(), with some-but-very-few tweaks !!!
!!! -> Maybe we could/should just edit ROME_ggtheme, but that might impact old code too... ???
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  p <- ggplot2::ggplot(mtcars, ggplot2::aes(x = factor(cyl))) +
    ggplot2::geom_bar() +
    ggplot2::labs(
      title = "A plot with non-OME-relevant data,",
      subtitle = "but which demos using theme_OME"
      ) +
   theme_OME()

  # Set as the global default for the current session:
  ggplot2::theme_set(theme_OME())
}

}
\author{
Dave Sirl
}
